requirements.txt

fastapi==0.115.6
uvicorn[standard]==0.30.6
sqlalchemy==2.0.32
pymysql==1.1.1
pydantic==2.8.2
python-multipart==0.0.9
jinja2==3.1.4
httpx==0.27.2
PyJWT==2.9.0
pytest==8.3.3
cryptography==42.0.8

